+++ tests/test_dcos_e2e/test_enterprise.py
@@ -4,17 +4,14 @@ Tests for using the test harness with a DC/OS Enterprise cluster.
 
 import json
 import subprocess
-import time
 import uuid
 from pathlib import Path
 
-import pytest
 import requests
 from passlib.hash import sha512_crypt
 
 from dcos_e2e.backends import ClusterBackend
 from dcos_e2e.cluster import Cluster
-from dcos_e2e.exceptions import DCOSTimeoutError
 from dcos_e2e.node import Role
 
 
@@ -399,19 +396,6 @@ class TestWaitForDCOS:
                 ip_detect_path=cluster_backend.ip_detect_path,
             )
             (master, ) = cluster.masters
-
-            for timeout_seconds in [1, 2]:
-                time_before_waiting = time.monotonic()
-                with pytest.raises(DCOSTimeoutError):
-                    cluster.wait_for_dcos_ee(
-                        superuser_username=superuser_username,
-                        superuser_password=superuser_password,
-                        timeout_seconds=timeout_seconds,
-                    )
-                time_after_waiting = time.monotonic()
-                time_difference = time_after_waiting - time_before_waiting
-                maximum_acceptable = timeout_seconds + 0.1
-                assert timeout_seconds < time_difference < maximum_acceptable
 
             cluster.wait_for_dcos_ee(
                 superuser_username=superuser_username,